+++ src/resources/zones/packages/index.ts
@@ -9,14 +9,5 @@ export {
   type PackageList,
   type PackageOutputBinding,
   type PackageSource,
-  type PackageRetrieveParams,
-  type PackageListParams,
-  type PackageRetrieveDraftParams,
 } from './packages';
-export {
-  Versions,
-  type PackageVersion,
-  type PackageVersionList,
-  type VersionRetrieveParams,
-  type VersionListParams,
-} from './versions';
+export { Versions, type PackageVersion, type PackageVersionList } from './versions';